Clarify breadcrumb params
The top-level `text` and `href` params should be used for the parent page, and not the current page.

This is because, as the guidance says, "You don't need to show the current page in the breadcrumb because this information is in the H1."

If you do not include the parent page as top-level params, but instead only within the `items` array, the result is identical on desktop views, but on mobile views will result in a back arrow but no text, and an empty link with the hidden text "Back to".