Merge outerContent and beforeContent blocks
This merges these two blocks into a single one named `beforeContent`, which appears within the `<div class="nhsuk-width-container">` container but before the `<main>` tag.

This should be used for components like the breadcrumb or back link, or any other component which should be within the same width as the rest of the page, but be skipped by the "skip to main content" link, not included when printing the page, and generally not considered part of the main content of the page.