Update back link position guidance
This updates the back link guidance to recommend that generally the back link should appear at the top-left of the page.

It also explicitly says that it should appear before the `<main>` tag in the HTML so that the skip link will skip past it.

Previously the guidance here suggested that some services included the back link at the bottom of the content. However both of the services mentioned seem to have since moved their back link to the top-left.