Fix to previous javascripts path
We made a small mistake in saying that the javascripts path in the tempaltes used to be `nhsuk-frontend.min.js` but it actually used to be `nhsuk.min.js`.

We've also realised that the path prefix is not necessarily `/javascripts/` but can vary depending on your setup (eg the NHS prototype kit uses `/nhsuk-frontend/`). So this makes that clearer by using `<JAVASCRIPTS-PATH>` to hopefully make it clear that that bit isn't part of the upgrade guidance.