Use camelCase for all input names
Generally we use camelCase for input names rather than hyphens, as it helps to distinguish the 2, and when used in the NHS Prototype kit or other javascript based frameworks, the camelCased properties can be accessed using the dot notation instead of square brackets (`data.fullName` instead of `data['full-name']`).

We didn't do this consistently though, so this switches all the remaining input examples to use camelCase.