Update the Text input examples and guidance
This updates the default Text input example to set the label as the page heading and to use the `large` font size. This makes it consistent with the other component examples.

Later down the page I’ve added some guidance on when to set the label as the page heading, with another example showing the label that is not the page heading.

Both examples also use a question phrasing for consistency with other examples.